requirements.txt
Django==6.0.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
gunicorn==23.0.0
dj-database-url==3.1.2
